backend/requirements.txt
asgiref==3.7.2
certifi==2023.7.22
distlib==0.3.7
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
filelock==3.12.3
pipenv==2023.9.1
platformdirs==3.10.0
pytz==2023.3
sqlparse==0.4.4
tzdata==2023.3
virtualenv==20.24.4
